requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.5
django-crispy-forms==1.14.0
django-environ==0.8.1
django-heroku==0.3.1
django-restframework==0.0.1
djangorestframework==3.13.1
gunicorn==20.1.0
mysqlclient==2.1.0
Pillow==9.1.1
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0
